requirements.txt
azure-ai-vision-imageanalysis
tqdm
python-dotenv
